% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXAdplyr.R
\name{src_exasol}
\alias{src_exasol}
\title{Create a dplyr connection to an EXASOL DB.}
\usage{
src_exasol(host, port = 8563, user = "sys", password = "exasol", ...,
  dsn = NULL, driver = NULL)
}
\description{
Create a dplyr connection to an EXASOL DB.
}
\section{Fields}{

\describe{
\item{\code{host}}{}

\item{\code{port}}{The port where the db listens.}

\item{\code{user}}{The DB user name. DEFAULT: 'sys'.}

\item{\code{password}}{The user's password. DEFAULT: 'exasol'.}

\item{\code{...}}{Further parameters, to be passed on to dbConnect.}

\item{\code{dsn}}{The ODBC dsn to connect to.}

\item{\code{driver}}{The ODBC driver library to use.}
}}

